made setup even better

+++ initial_setup.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
-echo "Running this application requires Spotify Client ID and Secret. Please enter them below:"
-read -p "Spotify Client ID: " SPOTIFY_CLIENT_ID
-read -p "Spotify Client Secret: " SPOTIFY_CLIENT_SECRET
-touch ./api/.spotify-env
-printf "SPOTIFY_CLIENT_ID=$SPOTIFY_CLIENT_ID\nSPOTIFY_CLIENT_SECRET=$SPOTIFY_CLIENT_SECRET" > ./api/.spotify-env
+if [ ! -f "./api/.spotify-env" ]; then
+  echo "Running this application requires Spotify Client ID and Secret. Please enter them below:"
+  read -p "Spotify Client ID: " SPOTIFY_CLIENT_ID
+  read -p "Spotify Client Secret: " SPOTIFY_CLIENT_SECRET
+  touch ./api/.spotify-env
+  printf "SPOTIFY_CLIENT_ID=$SPOTIFY_CLIENT_ID\nSPOTIFY_CLIENT_SECRET=$SPOTIFY_CLIENT_SECRET" > ./api/.spotify-env
+fi
 
 echo "Generating SSL certs for API and Website"
 cd ./api/ssl-certs
@@ -18,6 +20,7 @@ sleep 15
 cd ./web-scraper
 pwd
 ./add_scraped_data.sh
+echo "Added scraped data to database."
 cd ..
 pwd
 docker compose down
